+++ src/frolladaptive.c
@@ -1,37 +1,56 @@
 #include "data.table.h"
 
+#undef HASNFWARN
+#define HASNFWARN                                                                                                                                                   \
+{                                                                                                                                                                   \
+  ans->status = 2;                                                                                                                                                  \
+  snprintf(end(ans->message[2]), 500, _("%s: has.nf=FALSE used but non-finite values are present in input, use default has.nf=NA to avoid this warning"), __func__);\
+}
+
+#undef HASNFMSGRERUN
+#define HASNFMSGRERUN                                                                                                                  \
+{                                                                                                                                      \
+  snprintf(end(ans->message[0]), 500, _("%s: non-finite values are present in input, re-running with extra care for NFs\n"), __func__);\
+}
+
+#undef HASNFMSGEXACTPROPAGATED
+#define HASNFMSGEXACTPROPAGATED                                                                                                                                                  \
+{                                                                                                                                                                                \
+  snprintf(end(ans->message[0]), 500, _("%s: non-finite values are present in input, na.rm=FALSE and algo='exact' propagates NFs properply, no need to re-run\n"), __func__);\
+}
+
 /* rolling adaptive fun - router for fun and algo
  * rfun enum rollfun_t routes to rolling function
  * algo = 0: fast
  *   adding/removing in/out of sliding window of observations
  * algo = 1: exact
- *   recalculate whole fun for each observation, for mean roundoff correction is adjusted, also support for NaN and Inf
+ *   recalculate whole fun for each observation, for mean roundoff correction is adjusted
  */
-void frolladaptivefun(rollfun_t rfun, unsigned int algo, double *x, uint64_t nx, ans_t *ans, int *k, double fill, bool narm, int hasna, bool verbose) {
+void frolladaptivefun(rollfun_t rfun, unsigned int algo, double *x, uint64_t nx, ans_t *ans, int *k, double fill, bool narm, int hasnf, bool verbose) {
   double tic = 0;
   if (verbose)
     tic = omp_get_wtime();
   switch (rfun) {
   case MEAN :
     if (algo==0) {
-      frolladaptivemeanFast(x, nx, ans, k, fill, narm, hasna, verbose);
+      frolladaptivemeanFast(x, nx, ans, k, fill, narm, hasnf, verbose);
     } else if (algo==1) {
-      frolladaptivemeanExact(x, nx, ans, k, fill, narm, hasna, verbose);
+      frolladaptivemeanExact(x, nx, ans, k, fill, narm, hasnf, verbose);
     }
     break;
   case SUM :
     if (algo==0) {
-      frolladaptivesumFast(x, nx, ans, k, fill, narm, hasna, verbose);
+      frolladaptivesumFast(x, nx, ans, k, fill, narm, hasnf, verbose);
     } else if (algo==1) {
-      frolladaptivesumExact(x, nx, ans, k, fill, narm, hasna, verbose);
+      frolladaptivesumExact(x, nx, ans, k, fill, narm, hasnf, verbose);
     }
     break;
   case MAX :
     if (algo==0 && verbose) {
-      //frolladaptivemaxFast(x, nx, ans, k, fill, narm, hasna, verbose); // frolladaptivemaxFast does not exists as of now
+      //frolladaptivemaxFast(x, nx, ans, k, fill, narm, hasnf, verbose); // frolladaptivemaxFast does not exists as of now
       snprintf(end(ans->message[0]), 500, _("%s: algo %u not implemented, fall back to %u\n"), __func__, algo, (unsigned int) 1);
     }
-    frolladaptivemaxExact(x, nx, ans, k, fill, narm, hasna, verbose);
+    frolladaptivemaxExact(x, nx, ans, k, fill, narm, hasnf, verbose);
     break;
   default:
     error(_("Internal error: Unknown rfun value in froll: %d"), rfun); // #nocov
@@ -41,14 +60,14 @@ void frolladaptivefun(rollfun_t rfun, unsigned int algo, double *x, uint64_t nx,
 }
 
 /* fast rolling adaptive mean - fast
- * when no info on NA (hasNA argument) then assume no NAs run faster version
+ * when no info on NF (has.nf argument) then assume no NFs run faster version
  * adaptive rollmean implemented as cumsum first pass, then diff cumsum by indexes `i` to `i-k[i]`
- * if NAs detected re-run rollmean implemented as cumsum with NA support
+ * if NFs detected re-run rollmean implemented as cumsum with NF support
  */
-void frolladaptivemeanFast(double *x, uint64_t nx, ans_t *ans, int *k, double fill, bool narm, int hasna, bool verbose) {
+void frolladaptivemeanFast(double *x, uint64_t nx, ans_t *ans, int *k, double fill, bool narm, int hasnf, bool verbose) {
   if (verbose)
-    snprintf(end(ans->message[0]), 500, _("%s: running for input length %"PRIu64", hasna %d, narm %d\n"), "frolladaptivemeanFast", (uint64_t)nx, hasna, (int) narm);
-  bool truehasna = hasna>0;                                     // flag to re-run if NAs detected
+    snprintf(end(ans->message[0]), 500, _("%s: running for input length %"PRIu64", hasnf %d, narm %d\n"), "frolladaptivemeanFast", (uint64_t)nx, hasnf, (int) narm);
+  bool truehasnf = hasnf>0;                                     // flag to re-run if NAs detected
   long double w = 0.0;
   double *cs = malloc(nx*sizeof(double));                       // cumsum vector, same as double cs[nx] but no segfault
   if (!cs) {                                                    // # nocov start
@@ -57,12 +76,12 @@ void frolladaptivemeanFast(double *x, uint64_t nx, ans_t *ans, int *k, double fi
     free(cs);
     return;
   }                                                             // # nocov end
-  if (!truehasna) {
+  if (!truehasnf) {
     for (uint64_t i=0; i<nx; i++) {                             // loop on every observation to calculate cumsum only
       w += x[i];                                                // cumulate in long double
       cs[i] = (double) w;
     }
-    if (R_FINITE((double) w)) {                                 // no need to calc this if NAs detected as will re-calc all below in truehasna==1
+    if (R_FINITE((double) w)) {                                 // no need to calc this if NAs detected as will re-calc all below in truehasnf==1
       #pragma omp parallel for num_threads(getDTthreads(nx, true))
       for (uint64_t i=0; i<nx; i++) {                           // loop over observations to calculate final answer
         if (i+1 == k[i]) {
@@ -73,14 +92,13 @@ void frolladaptivemeanFast(double *x, uint64_t nx, ans_t *ans, int *k, double fi
           ans->dbl_v[i] = fill;                                 // position in a vector smaller than obs window width - partial window
         }
       }
-    } else {                                                    // update truehasna flag if NAs detected
-      if (verbose)
-        snprintf(end(ans->message[0]), 500, _("%s: non-finite values are present in input, re-running with extra care for NAs\n"), __func__);
-      w = 0.0;
-      truehasna = true;
+    } else {                                                    // update truehasnf flag if NAs detected
+      if (hasnf==-1) HASNFWARN
+      if (verbose) HASNFMSGRERUN
+      w = 0.0; truehasnf = true;
     }
   }
-  if (truehasna) {
+  if (truehasnf) {
     uint64_t nc = 0, pinf = 0, ninf = 0;                        // running NA counter
     uint64_t *cn = malloc(nx*sizeof(uint64_t));                 // cumulative NA counter, used the same way as cumsum, same as uint64_t cn[nx] but no segfault
     if (!cn) {                                                  // # nocov start
@@ -173,7 +191,7 @@ void frolladaptivemeanFast(double *x, uint64_t nx, ans_t *ans, int *k, double fi
       }
     }
     free(cninf); free(cpinf); free(cn);
-  } // end of truehasna
+  } // end of truehasnf
   free(cs);
 }
 /* fast rolling adaptive mean exact
@@ -181,14 +199,14 @@ void frolladaptivemeanFast(double *x, uint64_t nx, ans_t *ans, int *k, double fi
  * requires much more cpu
  * uses multiple cores
  */
-void frolladaptivemeanExact(double *x, uint64_t nx, ans_t *ans, int *k, double fill, bool narm, int hasna, bool verbose) {
+void frolladaptivemeanExact(double *x, uint64_t nx, ans_t *ans, int *k, double fill, bool narm, int hasnf, bool verbose) {
   if (verbose)
-    snprintf(end(ans->message[0]), 500, _("%s: running in parallel for input length %"PRIu64", hasna %d, narm %d\n"), "frolladaptivemeanExact", (uint64_t)nx, hasna, (int) narm);
-  bool truehasna = hasna>0;                                     // flag to re-run if NAs detected
-  if (!truehasna || !narm) {                                    // narm=FALSE handled here as NAs properly propagated in exact algo
+    snprintf(end(ans->message[0]), 500, _("%s: running in parallel for input length %"PRIu64", hasnf %d, narm %d\n"), "frolladaptivemeanExact", (uint64_t)nx, hasnf, (int) narm);
+  bool truehasnf = hasnf>0;                                     // flag to re-run if NAs detected
+  if (!truehasnf || !narm) {                                    // narm=FALSE handled here as NAs properly propagated in exact algo
     #pragma omp parallel for num_threads(getDTthreads(nx, true))
     for (uint64_t i=0; i<nx; i++) {                             // loop on every observation to produce final answer
-      if (narm && truehasna) {
+      if (narm && truehasnf) {
         continue;                                               // if NAs detected no point to continue
       }
       if (i+1 < k[i]) {
@@ -198,7 +216,7 @@ void frolladaptivemeanExact(double *x, uint64_t nx, ans_t *ans, int *k, double f
         for (int j=-k[i]+1; j<=0; j++) {                        // sub-loop on window width
           w += x[i+j];                                          // sum of window for particular observation
         }
-        if (R_FINITE((double) w)) {                             // no need to calc roundoff correction if NAs detected as will re-call all below in truehasna==1
+        if (R_FINITE((double) w)) {                             // no need to calc roundoff correction if NAs detected as will re-call all below in truehasnf==1
           long double res = w / k[i];                           // keep results as long double for intermediate processing
           long double err = 0.0;                                // roundoff corrector
           for (int j=-k[i]+1; j<=0; j++) {                      // sub-loop on window width
@@ -209,23 +227,21 @@ void frolladaptivemeanExact(double *x, uint64_t nx, ans_t *ans, int *k, double f
           if (!narm) {
             ans->dbl_v[i] = (double) w;
           }
-          truehasna = true;                                     // NAs detected for this window, set flag so rest of windows will not be re-run
+          truehasnf = true;                                     // NAs detected for this window, set flag so rest of windows will not be re-run
         } else {
           ans->dbl_v[i] = (double) w;                           // Inf and -Inf
         }
       }
     }
-    if (truehasna) {
+    if (truehasnf) {
+      if (hasnf==-1) HASNFWARN
       if (verbose) {
-        if (narm) {
-          snprintf(end(ans->message[0]), 500, _("%s: non-finite values are present in input, re-running with extra care for NAs\n"), __func__);
-        } else {
-          snprintf(end(ans->message[0]), 500, _("%s: non-finite values are present in input, na.rm was FALSE so in 'exact' implementation NAs were handled already, no need to re-run\n"), __func__);
-        }
+        if (narm) HASNFMSGRERUN
+        else HASNFMSGEXACTPROPAGATED
       }
     }
   }
-  if (truehasna && narm) {
+  if (truehasnf && narm) {
     #pragma omp parallel for num_threads(getDTthreads(nx, true))
     for (uint64_t i=0; i<nx; i++) {                             // loop over observations to produce final answer
       if (i+1 < k[i]) {
@@ -265,16 +281,16 @@ void frolladaptivemeanExact(double *x, uint64_t nx, ans_t *ans, int *k, double f
         }
       }
     }
-  } // end of truehasna
+  } // end of truehasnf
 }
 
 /* fast rolling adaptive sum - fast
  * same as adaptive mean fast
  */
-void frolladaptivesumFast(double *x, uint64_t nx, ans_t *ans, int *k, double fill, bool narm, int hasna, bool verbose) {
+void frolladaptivesumFast(double *x, uint64_t nx, ans_t *ans, int *k, double fill, bool narm, int hasnf, bool verbose) {
   if (verbose)
-    snprintf(end(ans->message[0]), 500, _("%s: running for input length %"PRIu64", hasna %d, narm %d\n"), "frolladaptivesumFast", (uint64_t)nx, hasna, (int) narm);
-  bool truehasna = hasna>0;
+    snprintf(end(ans->message[0]), 500, _("%s: running for input length %"PRIu64", hasnf %d, narm %d\n"), "frolladaptivesumFast", (uint64_t)nx, hasnf, (int) narm);
+  bool truehasnf = hasnf>0;
   long double w = 0.0;
   double *cs = malloc(nx*sizeof(double));
   if (!cs) {                                                    // # nocov start
@@ -283,7 +299,7 @@ void frolladaptivesumFast(double *x, uint64_t nx, ans_t *ans, int *k, double fil
     free(cs);
     return;
   }                                                             // # nocov end
-  if (!truehasna) {
+  if (!truehasnf) {
     for (uint64_t i=0; i<nx; i++) {
       w += x[i];
       cs[i] = (double) w;
@@ -300,13 +316,12 @@ void frolladaptivesumFast(double *x, uint64_t nx, ans_t *ans, int *k, double fil
         }
       }
     } else {
-      if (verbose)
-        snprintf(end(ans->message[0]), 500, _("%s: non-finite values are present in input, re-running with extra care for NAs\n"), __func__);
-      w = 0.0;
-      truehasna = true;
+      if (hasnf==-1) HASNFWARN
+      if (verbose) HASNFMSGRERUN
+      w = 0.0; truehasnf = true;
     }
   }
-  if (truehasna) {
+  if (truehasnf) {
     uint64_t nc = 0, pinf = 0, ninf = 0;                        // running NA counter
     uint64_t *cn = malloc(nx*sizeof(uint64_t));                 // cumulative NA counter, used the same way as cumsum, same as uint64_t cn[nx] but no segfault
     if (!cn) {                                                  // # nocov start
@@ -405,14 +420,14 @@ void frolladaptivesumFast(double *x, uint64_t nx, ans_t *ans, int *k, double fil
 /* fast rolling adaptive sum - exact
  * same as adaptive mean exact
  */
-void frolladaptivesumExact(double *x, uint64_t nx, ans_t *ans, int *k, double fill, bool narm, int hasna, bool verbose) {
+void frolladaptivesumExact(double *x, uint64_t nx, ans_t *ans, int *k, double fill, bool narm, int hasnf, bool verbose) {
   if (verbose)
-    snprintf(end(ans->message[0]), 500, _("%s: running in parallel for input length %"PRIu64", hasna %d, narm %d\n"), "frolladaptivesumExact", (uint64_t)nx, hasna, (int) narm);
-  bool truehasna = hasna>0;
-  if (!truehasna || !narm) {
+    snprintf(end(ans->message[0]), 500, _("%s: running in parallel for input length %"PRIu64", hasnf %d, narm %d\n"), "frolladaptivesumExact", (uint64_t)nx, hasnf, (int) narm);
+  bool truehasnf = hasnf>0;
+  if (!truehasnf || !narm) {
     #pragma omp parallel for num_threads(getDTthreads(nx, true))
     for (uint64_t i=0; i<nx; i++) {
-      if (narm && truehasna) {
+      if (narm && truehasnf) {
         continue;
       }
       if (i+1 < k[i]) {
@@ -428,23 +443,21 @@ void frolladaptivesumExact(double *x, uint64_t nx, ans_t *ans, int *k, double fi
           if (!narm) {
             ans->dbl_v[i] = (double) w;
           }
-          truehasna = true;                                     // NAs detected for this window, set flag so rest of windows will not be re-run
+          truehasnf = true;                                     // NAs detected for this window, set flag so rest of windows will not be re-run
         } else {
           ans->dbl_v[i] = (double) w;                           // Inf and -Inf
         }
       }
     }
-    if (truehasna) {
-      if (verbose) {
-        if (narm) {
-          snprintf(end(ans->message[0]), 500, _("%s: non-finite values are present in input, re-running with extra care for NAs\n"), __func__);
-        } else {
-          snprintf(end(ans->message[0]), 500, _("%s: non-finite values are present in input, na.rm was FALSE so in 'exact' implementation NAs were handled already, no need to re-run\n"), __func__);
-        }
+    if (truehasnf) {
+      if (hasnf==-1) HASNFWARN
+        if (verbose) {
+          if (narm) HASNFMSGRERUN
+          else HASNFMSGEXACTPROPAGATED
       }
     }
   }
-  if (truehasna && narm) {
+  if (truehasnf && narm) {
     #pragma omp parallel for num_threads(getDTthreads(nx, true))
     for (uint64_t i=0; i<nx; i++) {
       if (i+1 < k[i]) {
@@ -476,12 +489,12 @@ void frolladaptivesumExact(double *x, uint64_t nx, ans_t *ans, int *k, double fi
 }
 
 /* fast rolling adaptive max - exact
- * for hasNA=FALSE it will not detect if any NAs were in the input, therefore could produce incorrect result, well documented
+ * for has.nf=FALSE it will not detect if any NAs were in the input, therefore could produce incorrect result, well documented
  */
-void frolladaptivemaxExact(double *x, uint64_t nx, ans_t *ans, int *k, double fill, bool narm, int hasna, bool verbose) {
+void frolladaptivemaxExact(double *x, uint64_t nx, ans_t *ans, int *k, double fill, bool narm, int hasnf, bool verbose) {
   if (verbose)
-    snprintf(end(ans->message[0]), 500, _("%s: running in parallel for input length %"PRIu64", hasna %d, narm %d\n"), "frolladaptivemaxExact", (uint64_t)nx, hasna, (int) narm);
-  if (narm || hasna==-1) { // fastest we can get for adaptive max as there is no algo='fast', therefore we drop any NA checks when hasNA=FALSE
+    snprintf(end(ans->message[0]), 500, _("%s: running in parallel for input length %"PRIu64", hasnf %d, narm %d\n"), "frolladaptivemaxExact", (uint64_t)nx, hasnf, (int) narm);
+  if (narm || hasnf==-1) { // fastest we can get for adaptive max as there is no algo='fast', therefore we drop any NA checks when has.nf=FALSE
     #pragma omp parallel for num_threads(getDTthreads(nx, true))
     for (uint64_t i=0; i<nx; i++) {
       if (i+1 < k[i]) {
@@ -503,16 +516,16 @@ void frolladaptivemaxExact(double *x, uint64_t nx, ans_t *ans, int *k, double fi
       free(isnan);
       return;
     }                                                               // # nocov end
-    bool truehasna = hasna>0;
-    for (uint64_t i=0; i<nx; i++) { // no openmp as this shuld be very fast
+    bool truehasnf = hasnf>0;
+    for (uint64_t i=0; i<nx; i++) { // no openmp as this should be very fast
       if (ISNAN(x[i])) {
-        truehasna = true;
+        truehasnf = true;
         isnan[i] = true;
       } else {
         isnan[i] = false;
       }
     }
-    if (!truehasna) { // not found any NAs
+    if (!truehasnf) { // not found any NAs
       #pragma omp parallel for num_threads(getDTthreads(nx, true))
       for (uint64_t i=0; i<nx; i++) {
         if (i+1 < k[i]) {